--HG--
branch : 1.x

+++ src/Orchard.Web/Modules/Orchard.Packaging/Services/PackageInstaller.cs
@@ -90,57 +90,72 @@ namespace Orchard.Packaging.Services {
             // add the package to the project
             projectManager.AddPackageReference(package.Id, package.Version);
 
-            return new PackageInfo
-            {
+            return new PackageInfo {
                 ExtensionName = package.Title ?? package.Id,
                 ExtensionVersion = package.Version.ToString(),
-                ExtensionType = package.Id.StartsWith(PackagingSourceManager.ThemesPrefix) ? DefaultExtensionTypes.Theme : DefaultExtensionTypes.Module,
+                ExtensionType = package.Id.StartsWith(PackagingSourceManager.GetExtensionPrefix(DefaultExtensionTypes.Theme)) ? DefaultExtensionTypes.Theme : DefaultExtensionTypes.Module,
                 ExtensionPath = applicationPath
             };
         }
 
         public void Uninstall(string packageId, string applicationPath) {
-            // this logger is used to render NuGet's log on the notifier
-            var logger = new NugetLogger(_notifier);
-
             string solutionPath;
+            string extensionFullPath = string.Empty;
+
+            if (packageId.StartsWith(PackagingSourceManager.GetExtensionPrefix(DefaultExtensionTypes.Theme))) {
+                extensionFullPath = HostingEnvironment.MapPath("~/Themes/" + packageId.Substring(PackagingSourceManager.GetExtensionPrefix(DefaultExtensionTypes.Theme).Length));
+            } else if (packageId.StartsWith(PackagingSourceManager.GetExtensionPrefix(DefaultExtensionTypes.Module))) {
+                extensionFullPath = HostingEnvironment.MapPath("~/Modules/" + packageId.Substring(PackagingSourceManager.GetExtensionPrefix(DefaultExtensionTypes.Module).Length));
+            }
+
+            if (string.IsNullOrEmpty(extensionFullPath) ||
+                !Directory.Exists(extensionFullPath)) {
+
+                throw new OrchardException(T("Package not found: {0}", packageId));
+            }
+
             // if we can access the parent directory, and the solution is inside, NuGet-uninstall the package here
             if (TryGetSolutionPath(applicationPath, out solutionPath)) {
+
+                // this logger is used to render NuGet's log on the notifier
+                var logger = new NugetLogger(_notifier);
+
                 var installedPackagesPath = Path.Combine(solutionPath, PackagesPath);
                 var sourcePackageRepository = new LocalPackageRepository(installedPackagesPath);
-                var project = new FileBasedProjectSystem(applicationPath) { Logger = logger };
-                var projectManager = new ProjectManager(
-                    sourcePackageRepository,
-                    new DefaultPackagePathResolver(installedPackagesPath),
-                    project,
-                    new ExtensionReferenceRepository(project, sourcePackageRepository, _extensionManager)
-                    ) { Logger = logger };
 
-                // add the package to the project
-                projectManager.RemovePackageReference(packageId);
+                try {
+                    var project = new FileBasedProjectSystem(applicationPath) {Logger = logger};
+                    var projectManager = new ProjectManager(
+                        sourcePackageRepository,
+                        new DefaultPackagePathResolver(installedPackagesPath),
+                        project,
+                        new ExtensionReferenceRepository(project, sourcePackageRepository, _extensionManager)
+                        ) {Logger = logger};
 
-                var packageManager = new NuGetPackageManager(
-                    sourcePackageRepository,
-                    new DefaultPackagePathResolver(applicationPath),
-                    new PhysicalFileSystem(installedPackagesPath) { Logger = logger }
-                    ) { Logger = logger };
-
-                packageManager.UninstallPackage(packageId);
-            } else {
-                // otherwise delete the folder
-
-                string extensionPath = packageId.StartsWith(PackagingSourceManager.ThemesPrefix)
-                                           ? "~/Themes/" + packageId.Substring(PackagingSourceManager.ThemesPrefix.Length)
-                                           : "~/Modules/" + packageId.Substring(PackagingSourceManager.ThemesPrefix.Length);
-
-                string extensionFullPath = HostingEnvironment.MapPath(extensionPath);
-
-                if (Directory.Exists(extensionFullPath)) {
-                    Directory.Delete(extensionFullPath, true);
+                    // add the package to the project
+                    projectManager.RemovePackageReference(packageId);
                 }
-                else {
-                    throw new OrchardException(T("Package not found: ", packageId));
+                catch {
+                    // Uninstalling the package at the solution level failed
                 }
+
+                try {
+                    var packageManager = new NuGetPackageManager(
+                        sourcePackageRepository,
+                        new DefaultPackagePathResolver(applicationPath),
+                        new PhysicalFileSystem(installedPackagesPath) {Logger = logger}
+                        ) {Logger = logger};
+
+                    packageManager.UninstallPackage(packageId);
+                }
+                catch {
+                    // Package doesnt exist anymore
+                }
+            }
+
+            // If the package was not installed through nuget we still need to try to uninstall it by removing its directory
+            if(Directory.Exists(extensionFullPath)) {
+                Directory.Delete(extensionFullPath, true);
             }
         }
